Fix premature SSL termination in the SSL / TLS runtimes
The SSL / TLS runtime implementations for `shutdown_send` and
`shutdown_receive` were in fact tearing down the full-duplex connection
instead of shutting down only one part of it. In fact, we can't shut
down a single part of the connection in these runtimes, as the TLS
shutdown requires both endpoints to be involved in order to prevent
truncation attacks.

A previous symptom of this problem was fixed in
https://github.com/inhabitedtype/httpaf/pull/84, though it's clear now
that it was the wrong fix.