Fix the interaction between backpressure and fairness / yielding
In the case where a request handler yields too much to other streams,
there was a bug where http/af's request body implementation didn't have
a chance to surface buffered bytes to the application, and instead
buffered them forever.

The fix is to make sure that we flush the request body before trying to
compute the next read operation.